fix error and split debounce
